+++ lib/config/day.dart
@@ -1,0 +1,81 @@
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:xwordle/utils/utils.dart';
+
+/// This class represents a Wordle Day
+class WordleDay {
+  /// The word for the day
+  final String word;
+
+  /// Current index of the word
+  final int index;
+
+  /// The date of the day
+  final DateTime date;
+
+  /// Next day
+  DateTime get next => date.add(Duration(days: 1));
+
+  /// Constructor
+  WordleDay(
+    this.word,
+    this.index,
+    this.date, {
+    this.totalWinners = 0,
+    this.totalLosers = 0,
+    this.totalPlayed = 0,
+  });
+
+  /// Total winners
+  int totalWinners;
+
+  /// Total losers
+  int totalLosers;
+
+  /// Number of people who played
+  int totalPlayed;
+
+  /// Returns a WordleDay from a map
+  factory WordleDay.fromMap(Map<String, dynamic> map) {
+    return WordleDay(
+      map['word'],
+      map['index'],
+      DateTime.fromMillisecondsSinceEpoch(map['date'] * 1000),
+      totalWinners: map['totalWinners'] ?? 0,
+      totalLosers: map['totalLosers'] ?? 0,
+      totalPlayed: map['totalPlayed'] ?? 0,
+    );
+  }
+
+  /// Returns the formatted duration till the next wordle
+  String get formattedDurationTillNext {
+    return DateUtil.getFormattedDuration(next);
+  }
+
+  /// Update the word
+  WordleDay updateWord(String word) {
+    return WordleDay(word, index, date);
+  }
+
+  /// Save day to day.json
+  void save() {
+    final file = File('day.json');
+    if (!file.existsSync()) {
+      file.createSync();
+    }
+    file.writeAsStringSync(jsonEncode(toMap()));
+  }
+
+  /// Returns a map from the day
+  Map<String, dynamic> toMap() {
+    return {
+      'word': word,
+      'index': index,
+      'date': date.millisecondsSinceEpoch ~/ 1000,
+      'totalWinners': totalWinners,
+      'totalLosers': totalLosers,
+      'totalPlayed': totalPlayed,
+    };
+  }
+}