+++ lib/handlers/error.dart
@@ -1,0 +1,5 @@
+Future<void> errorHandler(Object err, StackTrace stack) {
+  print('Caught error: $err');
+  print(stack);
+  return Future.value();
+}