+++ lib/models/session.dart
@@ -1,8 +1,0 @@
-import 'package:televerse/televerse.dart';
-
-class WordleSession extends Session {
-  @override
-  Map<String, dynamic> toJson() {
-    throw UnimplementedError();
-  }
-}