+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 1.0.0
 
 - Initial version.
+- The Wordle game inspired by the [Wordle game](https://www.nytimes.com/games/wordle/index.html) from the New York Times as Telegram bot.
+- Every day a new word is generated.